Feature: Update Laravel AND Image Uploader

+++ src/config/jwt.php
@@ -2,5 +2,5 @@
 
 return [
     'secret' => env('JWT_SECRET', 'ukGKb5hXiBpPnPGYYH3NylUnELzeDAr6D54oax2VmecxDaLK3QG57dVLOFH7Jpoa'),
-    'ttl' => env('JWT_TTL', 86400), // 7 días
+    'ttl' => (int) env('JWT_TTL', 86400), // 7 días
 ];
